Move chat functions back to worldobject.
