+++ src/game/Objects/Unit.h
@@ -1072,7 +1072,7 @@ class MANGOS_DLL_SPEC Unit : public WorldObject
         uint32 getRaceMask() const { return getRace() ? 1 << (getRace()-1) : 0x0; }
         uint8 getClass() const { return GetByteValue(UNIT_FIELD_BYTES_0, 1); }
         uint32 getClassMask() const { return getClass() ? 1 << (getClass()-1) : 0x0; }
-        uint8 getGender() const { return GetByteValue(UNIT_FIELD_BYTES_0, 2); }
+        uint8 getGender() const override { return GetByteValue(UNIT_FIELD_BYTES_0, 2); }
 
         /**
          * @brief Inits display id to player display ID, depending on race, class, gender
@@ -1914,23 +1914,8 @@ class MANGOS_DLL_SPEC Unit : public WorldObject
         float getNativeScale() const;
         void setNativeScale(float scale);
 
-        void PMonsterSay(const char* text, ...);
-        void PMonsterSay(int32 text, ...) const;
-        void PMonsterYell(const char* text, ...);
-        void PMonsterYell(int32 text, ...) const;
-
-        void MonsterSay(const char* text, uint32 language = 0, Unit* target = nullptr);
-        void MonsterYell(const char* text, uint32 language = 0, Unit* target = nullptr);
-        void MonsterTextEmote(const char* text, Unit* target = nullptr, bool IsBossEmote = false);
-        void MonsterWhisper(const char* text, Unit* target = nullptr, bool IsBossWhisper = false) const;
-        void MonsterSay(int32 textId, uint32 language = 0, Unit* target = nullptr) const;
-        void MonsterYell(int32 textId, uint32 language = 0, Unit* target = nullptr) const;
-        void MonsterTextEmote(int32 textId, Unit* target = nullptr, bool IsBossEmote = false) const;
-        void MonsterWhisper(int32 textId, Unit* receiver, bool IsBossWhisper = false) const;
-        void MonsterYellToZone(int32 textId, uint32 language = 0, Unit* target = nullptr) const;
-        void MonsterScriptToZone(int32 textId, ChatMsg type, uint32 language = 0, Unit* target = nullptr) const;
-        static void BuildMonsterChat(WorldPacket *data, ObjectGuid senderGuid, uint8 msgtype, char const* text, uint32 language, char const* name, ObjectGuid targetGuid);
-
+        
+        
     protected:
         explicit Unit ();
